Split gateway types
