+++ packages/types/src/index.ts
@@ -2,5 +2,6 @@ export * from './discord.js'
 export * from './discord/reference.js'
 export * from './discord/webhookEvents.js'
 export * from './discord/interactions.js'
+export * from './discord/gateway.js'
 export * from './discordeno.js'
 export * from './shared.js'
